refactor: déplacer dashboard vers features/dashboard/
