+++ src/features/dashboard/Dashboard.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { DashboardView } from '@/views/DashboardView';
-import { useDashboard } from '@/hooks/useDashboard';
+import { DashboardView } from './DashboardView';
+import { useDashboard } from './useDashboard';
 
 export default function Dashboard() {
   const logic = useDashboard();
