+++ src/features/dashboard/__tests__/useDashboard.test.ts
@@ -1,7 +1,7 @@
 // src/__tests__/hooks/useDashboard.test.ts
 import { describe, it, expect, vi } from 'vitest';
 import { renderHook, waitFor } from '@testing-library/react';
-import { useDashboard } from '@/hooks/useDashboard';
+import { useDashboard } from '@/features/dashboard/useDashboard';
 import { createHookWrapper } from '@/__tests__/utils/test-utils';
 
 vi.mock('@/shared/hooks/useNavigationAdapter', () => ({
